[Gradle Build Model] Add a FAKE property type and throw when operations are unsupported
This change adds a FAKE PropertyType for fake element, we use this to throw
when attempting to perform an unsupported operation of a fake element.

For example it doesn't make sense to rename or convert a fake element to
a list of map. We throw UnsupportedOperationExceptions when this is
attempted.

Test: Test added and existing tests updated
Bug: N/A
Change-Id: I4a7a05df15ce8b55ef034d29e3bd49c8ff55b552