Handle operator overloading correctly in simpleperf
Previously, NodeNameParser was throwing an exception due to operator
overloading of type "<" or "<<". The cause is that the parser
tries to find matching character for ever opening bracket "<",
but operator overloading never has a matching one.

In addition to throwing an exception, the parser was not parsing
a method with operator overloading correctly.

Test: added unit tests in NodeNameParser:
      testCppOperatorOverloadingMethod,
      testCppOperatorOverloadingAndTemplatesInParameterAndInNamespace,
      testCppOperatorBoolOverloading
Change-Id: I7432d255b4e6f1019cd91f9d3e0d400b3dd60e88
Fixes: b/76088669