Updated Profilers loop to wait for preferred device
Make the device selection heuristics to be less smart about
selecting a device that has profilable processes. Previously,
the profilers always try to select a device that has online
processes if the current selected device is dead/does not have any.

The new logic does not switch devices automatically. Also, if a
preferred device has been set, the profilers will wait for that
device to come online and not select another one, unless the user
manually selects a different device. This new logic mirrors the
preferred process selection heuristics.

Bug: 77541508
Test: preferredDeviceHasPriority, updated existing tests as well
Change-Id: Iaf5fadf7df21c54e3d011bdb5df727d6ad45b56b