Fix GuiTestOptions.isRunningOnRelease
The idea.gui.test.remote.ide.path system property is only set on the
server, leading to incorrect results on the client, where it is used to
determine the path to the test data.

This change introduces a separate system property that will only be set
when running UI tests on the release.

Bug: 77588176
Change-Id: If98b0acf197e6216ccabb405828d96505fe2570b