Enable running tests with GuiTestRemoteRunner under Bazel
There were two main issues:
* The classpath argument exceeded OS argument length limits
* The configuration work done in IdeaTestSuiteBase only applies to the
server, and doesn't transfer automatically to the client, as it is a
separate process.

This change fixes these by:
* Creating a .jar that contains only a manifest, which specifies the
classpath. This jar is then used for the classpath argument.

* Copying the system properties settings from IdeaTestSuiteBase into
GuiTestLauncher. The static initializer in IdeaTestSuiteBase is already
run in the server process, so no additional work is needed there.

Also go back to old mechanism for finding gradle distribution, for now.

Test: Bazel UI test PSQ; also ran locally
Bug: 77655849
Change-Id: I073021d4f0940d29e20b93722da0da2874fbce90