Stop session should keep the session selected
Previously, stopping a session sets the device to null which
causes the profilers to think it should go to the NullMonitorStage.

The updated logic keeps the previous session when setting device to
null, and we only navigate to the NullMonitorStage iff the device is
set but the process is null.

Bug: 77604127
Test: updated
Change-Id: I6b037d9f2f1a1fb6111a7cfd5b4abf112d7011fd