Always select the capture in import trace mode
If the user attempts to unselect the capture, we will
automatically select the entire range of the capture.
CpuProfilerStage is a good place to handle logic specific
to "import trace mode", which promotes this solution.

One alternative is to keep the current selection range in
such scenarios, not selecting the entire capture range.
However, that would require SelectionModel or CaptureModel
to be aware of "import trace mode", which isn't ideal.

Fixes: 77204101

Test: CpuProfilerStageTest#captureAlwaysSelectedInImportTraceMode
Change-Id: I92106d5f6308217453d0a7b750ca6fff37c28f89