Use system property to specify ApplicationStarter
While this mechanism is only strictly necessary for running tests on
release bundles launched via native launchers, use it in the other cases
as well for consistency.

This CL also includes logic to create a VM options file that is
augmented with options to set the application starter and server port.
When launching a "remote" client IDE, the STUDIO_VM_OPTIONS environment
variable is set to point to this file.

Bug: 77820421
Test: RestartIdeTest
Test-Strategy: ui.test.groups=TEST_FRAMEWORK
Change-Id: Ia9fb56093632aecb750d100b2902e4154cc1a149