Dynamic Apps: Deploy flat APK for API level < 21
When an Android application has at least one dynamic feature, the build
output is usually split APKs. However, on pre-L devices, split apks are
not supported (i.e. adb install-multiple does not work).

In such a configuration, Android Studio must fall back to using the
bundle tool (as opposed to the regular "assemble" task) to create
the build output. The bundle tool recognizes the target API does not
support split apks, so it will create a "flat" APK (i.e. a single APK
that contains the base apk and all the features).

This CL implements this automatic fallback mechanism, so it is transparent
to the end user.

Bug: 73074662
Test: Included
Change-Id: I3514b1a7e4ad7322c016a117985b12057a53e84f