Support breakpoint in interface methods with desugaring
Non-abstract methods in interfaces are allowed starting with Android
API 24. In order to be compatible with prior versions of Android, an
application needs to run a desugaring phase which includes
transforming interfaces with non-abstract methods.

After desugaring, the only non-abstract method in an interface is the
<clinit> method. Other methods (default or static) have been moved to
a companion class synthesized by the desugaring phase. Therefore,
setting a breakpoint in one of these methods in the interface will no
longer works because there is no code at runtime.

This CL modifies the dedicated position manager for Android to look
for the position of the breakpoint in any inner class of the
interface. Therefore, if a companion class is created, the debugger
will take it into account to set the breakpoint.

Bug: 62286632
Bug: 77235562
Test: add unit tests in InstantRunPositionManagerTest
Change-Id: I47fda8293a6b79bbc3af208e747eefbef8c4ae2e