Updating AtraceDecompressor to trim line to 1023 characters.
The reason for this limit is due to a bug in trebuchet. We want to
trim the characters to below the maxline length defined in trebuchet
to prevent the message from being dropped. If the message is dropped
then we end up with mis-matched begin and end events causing our
results to look different than systrace.
Systrace handles long lines by truncating them as such we are modeling
that behavior. The difference between us and systrace is
we count the entire line vs systrace counts only the message. As such
we will truncate slightly more data than systrace, and our data will be
variable depending on thread name, thread id, and process id.

Test: AtraceDecompressor updated to test long lines.
Change-Id: I2eff924610e5ea8b8889559ffeb1512fdb98023b
Fixes: 77846431