Update capture/profiler state when parsing a capture
We're currently setting the capture state to PARSING when parsing the
captures for the first time. We should also should make sure to set
the state when parsing the captures again, which happens, for instance,
when closing and reopening CPU profiler.

We should also make sure to restore the previous state after PARSING a
capture. For instance, if we select a capture (which then gets parsed)
while there is a capture in progress, we should go back to CAPTURING
state when finish parsing.

Another thing we should do is to set the ProfilerMode to EXPANDED when
start parsing. This way, we don't see the height of the capture details
panel grow suddenly after the capture is parsed.

Test: All tests added/modified are in CpuProfilerStageTest. Modified
      captureSuccessfully to include a check for ProfilerMode#EXPANDED
      and updated testElapsedTime to reflect the changes made. Also added
      setCaptureWhileCapturingShouldParseAndReturnToCapturing,
      setCaptureWhileIdleShouldParseAndReturnToIdle, and
      getCaptureFutureShouldSetStateToParsing.

Bug: 75955739

Change-Id: Ife504ab57f3efc496ae15d8f9e627de68cb27f3d