Use AVD API 24 in AbiSplitApksTest
As observed on tests run history, the testX64AbiSplitApks test is very
flaky while testX86AbiSplitApks is stable. As informed that for API 24
on Linux, the cold booting time of x86_64 is slightly slower than x86
(~2 seconds) on Linux machine. This slowness could be doubled or tripped
on build bot slave machine. (We don't have measurements for API 23).
Using AVD API 24 other than API 23 in test, and will monitor the future
test results on build bot

bug: 76413115
Test: Ran AbiSplitApksTest tests locally, passed
Change-Id: I98d924e65a03cfdb0ff106424a7907f90bde002b