Wait for sync to start when testing in-progress sync
NlEditorTest#designEditorUnavailableIfInProgressBazelSyncFails ensures
that the layout editor doesn't render when a layout file is opened
during a sync that eventually fails. This test has been failing
consistently because it hasn't been waiting for the project sync to
start before opening the layout file.

This CL resolves the issue by introducing a waitForProjectSyncToStart
method to GuiTestProjectSystem. The Bazel implementation of this method
relies on the content of the Bazel console to determine when a project
sync has begun. BazelConsoleToolWindowFixture#hasSyncStarted has also
been updated to look for the notification that an incremental project
sync has begun (previously it was looking for any sync task, which
caused it to return true prematurely).

Bug: 77974117
Test: Ran designEditorUnavailableIfInProgressBazelSyncFails 50 times
      locally with no flakes (before I was seeing failures every other
      run).

Change-Id: Idfd711abf26661f939c976b32fddd2384b97cd51