Test instant apps are deployed to Android systems
InstantAppRunTest.createAndRunInstantApp was recently updated to avoid
using a real emulator when testing to minimize test flakiness due to
variance in AVD boot times. That test is now CreateAndRunInstantApp.

However, since the test no longer uses the real emulator, there is a gap
in testing where deploying instant apps to the emulator is not tested.
InstantAppTest covers that gap by testing that APKs can be deployed to a real
emulator, but the test does so without using a UI, which is already
covered by CreateAndRunInstantApp.

Bug: 77872027
Test: InstantAppTest and CreateAndRunInstantApp
Change-Id: I61cc95ee2fd5c4ca3809d8380a56d6cc60bd8579