Prevent multiple callstacks from being selected at the same time
Some callstack views are logically related (e.g. multiple callstacks
in a vertical list), so when one is selected, the other's selection
should be cleared.

To handle this, we introduce the concept of a StackTraceGroup, which
is a sort of factory for creating StackTraceViews, except the views
are associated with each other so only one can be actively selected
at a time.

Also removed some unecessary public methods from APIs due to this
change.

Change-Id: I09521fc0f253eb72e1f5d0b9fc32aeb34579e8ce
Fixes: 77877890
Test: Added