CPU config dropdown is always the ongoing profiling
Previously, the dropdown always was resetting to "Sampled Java",
which is confusing especially when the user has an ongoing profiling
and the selected config in dropdown is not same as ongoing profiling
config.

Also, this change improves UX experience for startup profiling as well,
because it was always showing "Sampled Java" too, regardless of the
actual selected CPU profiler configuration of startup profiling.

Test: added a unit test in CpuProfilerStageTest:
      configurationShouldBeTheOnGoingProfilingAfterExitAndEnter
Bug: 77592530
Change-Id: Iad616c343bb0c347b950f3483dfc5234716c22fa