Add getResolvedDependency to AndroidModuleSystem.
Project system's dependency management APIs were designed with the
assumption of a well defined closure of dependencies that will satisfy
the needs of the IDE.  Dependencies are defined as enums and the API
doesn't allow creation of custom dependencies.  As the list of
dependencies needed by Android Studio grow, we can't always predict what
dependencies we need.

This CL is the second of a set of CLs that reimplement project system's
dependency management API to work with custom dependencies. Users can
now use maven coordinates to specify custom dependencies and still
remain build system independent.

This CL adds getResolvedDependency which is analogous to
getResolvedVersion. The dependsOn function in DependencyManagement.kt
has been changed to use this new API as an example usage.

The old API is kept to preserve existing implementations until this CL
goes in.  A future CL will migrate existing usages to the new API and
remove the old API.

Bug: 77874749
Test: Added new tests to GradleModuleSystemIntegrationTest.
Change-Id: I273d37c9aa83d238c3fbe702a76ba7569eee0c2d