[New Dynamic Module] Add Activities to new Module
When adding a new Activity to a Dynamic Module, all resources
referenced in the Manifest must be copied to the Base Feature.
This is a limitation of the Build System that requires that all
the values of the merged manifest are in the Base. This is also
similar to Instant Apps.

Bug: 78512556
Test-Strategy: ui.test.groups=PROJECT_WIZARD
Test: Added AddDynamicFeatureTest@addLoginActivityToDynamicModule
Change-Id: I6fdb4152448168dab4bf5b00b5647e934feb606b