Build threads model before setting an imported capture
We were creating the threads model only after setting a capture in
import trace mode.

There is no problem with that in terms of model, as we get to the
current state when both capture and threads model are set. However,
setting the capture also updates the threadId in CpuCaptureModel and, by
the time we try to select the thread ID corresponding to the main thread,
CpuCaptureModel will think we're trying to set the same thread and won't
fire the aspect to update the UI.

Changing the order and building the threads model before setting the
capture will cause the aspect to be properly fired when setting the
capture and consequently make the main thread to be selected in the UI.

Bug: 78191180
Test: Updated ImportedTraceThreadDataSeriesTest tests. UI response can
      be tested manually by importing a trace file and verifying the
      main thread gets selected.

Change-Id: I98c46b9f2dd38490313fbf2d781203f3ff636dc2