Removes StateChartModel aspect dependency from updater
The StateChartModel was getting update ticks from the Updater,
causing the transforms to continuously get recomputed. Now the
StateChartModel registers itself to the Range's aspect to
receive updates.

Also fixes a few more double/float issues not caught in the
previous commit.

Also fixes an issue where the antialiasing setting was set to
the wrong RGB order, causing slowdowns on repaint.

Bug: 77151915
Change-Id: If692d752997b28116047682d425dca1082ae13a9
Test: Updated tests to compensate for this change.