Add getAvailableDependency and registerDependency
Project system's current dependency management APIs were designed
assuming a well defined closure of dependencies that will satisfy
all the needs of the IDE.  Because of this, dependencies are defined
as enums and the API doesn't allow creation of custom dependencies.
However as the list of needed dependencies grow, we can't always
predict what dependencies we need.

This CL is the third of a set of CLs that reimplement project system's
dependency management API to work with custom dependencies. Users can
now use maven coordinates to specify custom dependencies and still
remain build system independent.

This CL adds getAvailableDependency to AndroidProjectSystem,
registerDependency to AndroidModuleSystem, and makes the following
changes:

* DependencyManagement.kt has been retrofitted with the new dependency
management APIs introduced in these set of CLs. This should completely
eliminate DependencyManagement.kt's reliance on the old APIs.

* Various test-purpose project system implementations have been
changed to prefer the usage of GradleCoordinates instead of
GoogleMavenArtifactIds.

* There was a minor bug in GoogleMavenArtifactIds that caused
artifact revision to be parsed incorrectly.  This is now fixed.

* New test methods are added to GradleModuleSystemIntegrationTest to
cover registerDependency and GradleProjectSystemIntegrationTest has
been created to cover getAvailableDependency.

* getDeclaredDependency has been renamed to getRegisteredDependency.

Bug: 77874749
Test: Added more tests to GradleModuleSystemIntegrationTest and created
GradleProjectSystemIntegrationTest.
Change-Id: If04c52bb64a1caf0394d88b152d641977ddda02f