Add duration to Session and update loading icon/text
This change improves the way SessionsManager syncs with the UI
to support updating/animating the UI components in place. We now
only refresh the sessions list if we detect changes in the proto
objects that back the SessionArtifacts.

Sessions now show durations. And ongoing heap dump and cpu captures
are represented with a spinning icon and "capturing..." text.

Bug: 75962959
Bug: 74975946
Test: added

Change-Id: I28e5abce3dc077e6d555432ec8103a2f8162b2b8