Fixes profilers' title sync'd to the previous session
Previously, the session and session meta data we retrieve
from StudioProfilers still point to the previous session
while we are in the middle of the aspect change update. This
CL makes sure we are reference the new value which lives in
SessionsManager.

We will better clean up how these are stored to two places
when we can completely remove the sessions flag.

Bug: 78251101
Test: updated
Change-Id: I7e82ba9c31da023fea57147d020a01a2209d5bbc