[Gradle Build Model] Move plugins to use GradlePropertyModel
This change makes plugins use GradlePropertyModels. In order to
do this we change the structure of the GradleDslElement tree to
keep information about the whole element the plugin was applied in.

We also create a new PluginModel which uses either a map or a literal
to represent each plugin entry.

Test: unit tests added/updated
Bug: N/A
Change-Id: I2ddb4c63b4865c5462243ea0891d5a69bb4a52b5