Fix CPU artifact subtitle for ongoing and imported captures
In the sessions panel, we usually display the capture start timestamp
relative to the session start time in CpuCaptureSessionArtifact. For
imported captures, we should display the file creation time instead,
otherwise we would always display 00:00:00 as a timestamp. If getting
the file properties fails, we should fallback to the session epoch start
time.

Ongoing captures also shouldn't display the capture approximated start
time, but "Capturing..." instead.

Bug: 78191180
Bug: 74975946
Test: Updated SessionsViewTest#testCpuOngoingCaptureItemSelection, and
      SessionsViewTest#testCpuCaptureItemSelection. Also added
      CpuProfilerTest#importedSessionsStartTimeShouldBeTraceCreationTime.
Change-Id: Ic2564aa1a0b096d1b3baf7133b154227293d5cef