Fix exception from data binding attribute.
Data binding allows specifying data bindings in XML which creates
various listeners. These attributes are not specified in attrs.xml
and thus the property inspector does not have the sufficient
information to handle these.
There is also some suspicious code in TagSnapshot.getValue
which blocks the read of such attribute values.
For now: disable these attributes in the properties panel.

Bug: 77234265
Test: Added unit test.
Change-Id: I7279a0b1cc2558a0c520b45ca0f796179129b3b1