Fix skin image for Nexus 7 (2013) tablet
The emulator does not display the skin correctly for the
Nexus 7 tablet. Although the WEBP image looks OK in Chrome,
there is something the emulator does not like about it.

The emulator is OK with the original PNG file. I used Studio to
re-convert that PNG to WEBP. The emulator is happy with this new
WEBP version.

Change-Id: I596d1644c2343841cda4db0b6bb08ba486e2ca37
Fixes: 71813076
Test: None added