Explicitly set CC envar for Bazel UI tests
There's an issue with running UI tests targeting Bazel projects on
some buildbots where Bazel sync isn't able to determine the location
of gcc, causing initial project syncs (and subsequently the tests)
to fail. To be more consistent, we should rely on the version of gcc
stored in prebuilts for our UI tests.

This CL resolves the issue by having the prepareTestForImport method
in BazelGuiTestProjectSystem explicitly set the CC envar in the .bazelrc
of the test workspace so that it points to the prebuilt gcc binary.

Bug: 73358558
Test: Manually tested by deleting the gcc symlink in my /usr/bin
      (which reproduces the issue) and verifying that
      BazelStartupSyncTest now passes.
Test-Strategy: ui.test.groups=DEFAULT,EDITING

Change-Id: I138ebc61d759bcd031740f230c0d92e9596176e0