Propagate exceptions thrown during creation of a RenderTask
Currently, all exceptions generated during the creation of a RenderTask
are swallowed with an incorrect assumption that such exceptions are only
possible when the project is being closed. However, the logs from
b/78303029 show that there may be exceptions even otherwise. This CL just
 propagates whatever exception was originally thrown unless the project is
 being closed.

Bug: 78303029
Test: Manually confirmed that we see the original exception in the logs
if one was thrown during the creation of a RenderTask

Change-Id: I86dde8aa302d26f5570f20fe856a84c24c7fd86d