Consolidate code that gets config path in UI tests
A second method for doing this came in with the JetBrains fork of the UI
test framework, and it used a different location than we had been using.

This CL moves the existing logic into GuiTests (in preparation for
possibly eventually deleting IdeTestApplication) and deletes the
duplicate implementation from GuiTestOptions.

Bug: 78237890
Change-Id: I8283749f064fa6d353d4d58043bc292cfe627377
Test-Strategy: ui.test.groups=TEST_FRAMEWORK,THEME