CpuProfilerStage should always refer to the same session
Everytime we needed a session in CpuProfilerStage, we were querying
StudioProfilers#getSession, which is fine for the general case. However,
there are some background tasks that reference the session at a later
point and that could result in a inconsistent behavior if the users
change the session before the task completes.

For example, importing a CPU trace will parse it in background and add a
capture to the session once parsing is done. If the user changes the
session while parsing is in progress, the capture will be set to the
newly selected session instead of the one created to accomodate the
capture itself.

Bug: 78384584
Test: CpuProfilerStageTest#sessionChangeShouldntAffectStageSession
Change-Id: Ideaa9f17a040b806df489ac5e509926922ddb7d0