Add support for tabs in the PropertiesPanel.
Add the concept of a named PropertiesTab which holds the
specification of the properties shown on any particular tab in the
properties panel.

A tab may be searchable or not. If a tab is not searchable that tab
will be removed from the visible list of tabs while searching in the
properties.
For Nele we want to hide the inspector tab and display only perform
searches in the advanced tab.

This is in preparation for implementing the advanced tab.

Bug: 73778668
Test: Added test for view switching and filtering.
Change-Id: I0a2771f2d31d2f79bedf985772b0deda38fab179