Fixing issue where zooming/panning would change view of capture.
If you zoomed in on an Atrace capture the thread states could change
depending on if the start of the capture was in view or not.
If the capture start is not in view, no sampled data would be requested.
This leads to the capture showing no data for areas that did not have
activity.
If the start was in view the capture would show the last sampled state
as the state up to the point of the first atrace state change. This
change addresses this issue by changing the query to query from
the start of the requested range to the first capture state. This
gives us sampled values up to the first state change from atrace.

Test: Updated MergeCaptureDataSeriesTest to check that we pass the
proper ranges.
Bug: None

Change-Id: I12d6b56ea5a0773b220b32089c48566c78f9c2fe