Adjusting sizing for CpuKernel / ThreadState charts.
The default height gap for state chart is 0.5, this is not the design
for threads / kernel view anymore as such we set it to 0.
I also added a bug for tracking the removal of StateChartConfig in
favor of a builder pattern for StateChart since we don't actually use
the config anywhere.

Test: None, added bug for tracking (b/78585673)
Bug: None, none filed yet.
Change-Id: Idf04f9b20d6389936fb19b9fa47d0131a35670b6