Profile processes detected after the Run/Debug/Profile actions
Previously, when user presses Run/Debug/Profile, if there is an existing
process with the same name running, the profiler can start profiling it
before the process is killed and relaunched.

This change forces the profiler to start profiling preferred processes
only if they appear after the time the profiler's launch task is performed.

Bug: 76426618
Bug: 78139887
Test: Updated StudioProfilersTest
Change-Id: I7d908205e70c8c549fa94d5b034eb31deed997e6