Support launching instant app with no specified URL
The new Instant App SDK library (used by Studio if present) supports
launching instant apps with no URLs. It will first look for a default
URL if the URL is blank, then it will launch the main activity of the
app using the new launch API.

Ideally we'd like this to let you choose between "Default Activity" and
"URL" in the launch options for instant apps, just as you would for
installable apps. However, we ran into some strange bugs with merged
manifest activities when pursuing that path, so for the moment we're
just letting you leave the URL blank if you are launching an instant app
and using the new SDK lib.

This is an important feature for I/O or just after since we're announcing
the new launch API.

bug: 78912753
Test: Tested manually + unit tests. Oddly the class DeepLinkLaunch doesn't have
tests to begin with.

Change-Id: I7bc04a3af5be8d8529144a6de2d4bffb48f53721