Fix resizing images issue
When painting to the screen in the ScreenViewLayer, in some platforms,
we would use an image that wasn't fully compatible with the screen.
Because of that, Swing would make an additional copy of a potentially
massive image.
This CL ensures that the image copy is compatible with the screen
so the paint does not involve a copy.

Test: N/A
Change-Id: I9b94b79099936a62c3a3f06f16aff8dccd0baf1a
Fixes: 78254948