Fix exception when opening CPU profiler in HiDPI
We were throwing an exception when opening CPU profiler level 2 in HiDPI
mode and/or using big font sizes (e.g. 24 or 36).

The reason for that is we were using TitledSeparator in the profiling
configurations combobox, and trying to manually revert the horizontal
border added by this class. The problem is TitledSeparator sets the
border insets without scaling the values, while we were trying to revert
them using JBUI.Borders.empty(...), which does the scaling. That was
causing our component to have negative width and, therefore, throwing an
exception.

We shouldn't use TitledSeparator for that purpose, as it's used for
named separators (e.g. "MySeparator -------"). Also, trying to revert
its logic is a bad practice, as we end up relying on implementation
details instead of an API.

Instead, we should simply use a horizontal JSeparator. TitledSeparator,
for instance, uses a JSeparator (and a JBLabel to display the separator
name, which we don't need).

Bug: 78913602
Test: Manual. Open CPU profiler using a HiDPI monitor and an exception
      shouldn't be thrown. The separator in the profiling configurations
      combobox should be displayed normally.

Change-Id: I3feca1a405d4ffdf8694617aa7f36a5c0859bfcc