[New PSD] Filter variable by reparsing.
Currently we populate the property editor drop-down list with all the
variables whose current values can be directly converted to values of
the type of the property. This approach often results in too many
irrelvant variables included on the list.

This CL changes filtering to only include those variables that are
accepted as text by the property.

Bug: n/a
Test: existing updated
Change-Id: Id210ca6587f0216f642a86f5372ebf4a95fe98da