Remove getResolvedVersion from AndroidModuleSystem
This CL removed all usages of getResolvedVersion and replaces it with
getResolvedDependency where applicable. NlPropertiesTest and
MockAppCompat has been changed to use actual gradle coordinates with
Mockito.

Note: A workaround placed in MockAppCompat due to Mockito's explicit
return of a null argument into a "when()" clause when, in kotlin, does
not accept null arguments. The workaround function argThat_NonNull
simply does what it should do in java.

This CL is a part of the migration effort from old dependency management
APIs to the newer, more flexible ones.  This CL only does in-place swaps
and does not make logical changes unless stated above.

Bug: 77874749
Test: Dependency management related tests pass.
Change-Id: I4985822bdca8fec40ec9ea065b90a32a80982f31