Remove polling logic from HttpDataFetcher
This feature originally intended to help refresh the UI when
it saw that a request switched from "downloading" to "finished
downloaded". This was something we used to see a lot more when
the network profiler was first launched, since due to poorly
optimized perfd code, requests with large payloads would hang,
leaving selections full of in-progress requests that had since
updated but the UI didn't realize it yet.

Now, however, the feature doesn't really help anymore, and can
definitely cause more harm than good. As long as there is a
single request in the connection table which never finishes
downloading (e.g. a request that was opened and never closed),
the fetcher will repeatedly trigger listeners that refresh the
UI and can cause a bad user experience by resetting a user's
selection underneath them.

Change-Id: I27d5ea271359f1ff0e42855b090c23e7d3046544
Fixes: 79222163
Test: Code deleted; all logic covered by existing tests