Style titles with dark background.
Instead of adding a horizontal separator line to make the title
lines stand out, this change is an attempt to implement the
design from csince@ where each title is bold text on a dark gray
background, and the background has border lines top & bottom.
An blank separator is also added between the title and the
elements in the inspector.

When the titles are collapsed, the spacing should be hidden and
each title should only have 1 border line to separate the titles.

Also: Remove the separator lines from the inspector, since they
served the same purpose before this change.

Bug: 73778668
Test: Tests are updated to match the separator change.
Change-Id: Ifdd12872a089495e3cbd352e7ad0d2156147f53f