Primary pieces of proto AAR support in Studio
There are still some places in Studio, particularly in Android lint,
that need to be retrofitted to support resources in res.apk, but most
of Studio functionality should work with proto AARs as a result of this
change.

Will try to run Studio with proto AARs with Chris' help later this week
or early next week.

Test: FileResourceOpenerTest, ResourceHelperTest
Bug: 74425399
Change-Id: I38e0ce2f13e6dc384cc5a85d58df910a0f9df551