Introduce IDE restart policies for UI tests
This change enables setting the policy for when to restart the IDE while
running UI tests via the idea.gui.test.restart.policy system property.
Options are EACH_TEST, TEST_FAILURE and IDE_ERROR, defaulting to
IDE_ERROR.

It also consolidates the IDE restart logic into the RunListener in
JUnitServerImpl. Previously IDE errors were checked (and a restart
triggered) at the beginning of a test, to see if the previous test had
errors. Now this is done at the end of a test.

Test: verified manually
Change-Id: Ied8d67ff470d647fc2ad11f453c1c7cfb3ad44c0
Test-Stragety: ui.test.groups=TEST_FRAMEWORK,THEME