Run AddKotlinTest within Bazel
There is a bug encountered while working on this commit. When there is
no network connection, the Android Studio Kotlin plugin uses version
1.0.0 of the Kotlin Gradle plugin when configuring Kotlin. When there is
a network connection, the Android Studio Kotlin plugin chooses version
1.2.41 of the Kotlin Gradle plugin. This test uses a workaround to
explicitly set the Kotlin Gradle plugin version to use in the root
Gradle file.

This commit also adds the //tools/idea:platform-build-info target as a
dependency, and this commit also adds the tools/idea/build.txt as a
symlink at the start of the test. The target and the file are needed in
order for the Kotlin plugin to be installed during the IDE's startup.

Fixes: 79776101
Bug: 79552030
Test: bazel test //tools/adt/idea/android-uitests:AddKotlinTest
Change-Id: I52a7294adfa71c148d7873c5ad6f860aa9583783