Major tooltip overhaul
This CL changes the parenting behavior of the tooltips
to be absolute to the ProfilerLayeredPane. It removes
the reparenting logic and moves all bounds calculations
to the layout. It also now properly calls revalidate
and repaint, instead of just invalidate.

This CL also fixes layout issues with CPU tooltips, as
those were not expanding the tooltip when needed.

NOTE: if you see anymore bugs with tooltips (especially
if the tooltip is too narrow), then it's most likely
the layout manager for that tooltip view is a
TabularLayout with "*" as the column sizer, instead of
"Fit".

Fixes: 79479019
Fixes: 79884359
Fixes: 78583667
Fixes: 71506019
Test: Fixed tests to pass

Change-Id: I9911770b58201b1ac8cc95845c275a53777190e5