Remove AndroidModuleSystem#addDependencyWithoutSync
This CL removes existing usages addDependencyWithoutSync and replaces it
with the new dependency management APIs where applicable. In particular:

* HtmlLinkManager has been modified to use DependencyManagent.kt instead
of the underlying project system APIs.

* HtmlLinkManagerTest has been changed from a TestCase to an
IdeaTestCase to allow the usage of TestProjectSystem. This removes the
need to mock out a copy of AndroidModuleSystem and greatly simplifies
the test.

This CL is a part of the migration effort from old dependency management
APIs to the newer, more flexible ones.  This CL only does in-place swaps
and does not make logical changes unless stated above.

Bug: 77874749
Test: Updated existing dependency management tests.
Change-Id: I8e97b6c437144e1f34e3d02cb34d324a5cff2958