[Gradle Build Model] Introduce RawText and ensure correct types after setting
This change introduces a RawText class which has behaviour similar to
ReferenceTo but allow you to provide intent that what you are
setting is not a reference but a string that should be written to the
property as is.

To do this we also correct the types of properties before applying and
reparsing.

Test: Unit tests added
Bug: N/A
Change-Id: Ic4964cf2fbb577a52b9b00f720dc38addcaf2195